优化

+++ src/main/java/com/sz/common/util/Mas.java
@@ -7,79 +7,79 @@ import java.math.BigDecimal;
  */
 public class Mas {
 	
-	public static double jia(double v1,double v2){
+	public static BigDecimal jia(double v1,double v2){
 		BigDecimal b1 = new BigDecimal(Double.toString(v1));
         BigDecimal b2 = new BigDecimal(Double.toString(v2));
-        return b1.add(b2).doubleValue();
+        return b1.add(b2);
 	}
 	
-	public static double jia(String v1,String v2){
+	public static BigDecimal jia(String v1,String v2){
 		BigDecimal b1 = new BigDecimal(v1);
         BigDecimal b2 = new BigDecimal(v2);
-        return b1.add(b2).doubleValue();
+        return b1.add(b2);
 	}
 	
-	public static double jia(BigDecimal v1,BigDecimal v2){
-        return v1.add(v2).doubleValue();
+	public static BigDecimal jia(BigDecimal v1,BigDecimal v2){
+        return v1.add(v2);
 	}
 	
-	public static double jian(double v1,double v2){
+	public static BigDecimal jian(double v1,double v2){
 		BigDecimal b1 = new BigDecimal(Double.toString(v1));
         BigDecimal b2 = new BigDecimal(Double.toString(v2));
-        return b1.subtract(b2).doubleValue();
+        return b1.subtract(b2);
 	}
 	
-	public static double jian(String v1,String v2){
+	public static BigDecimal jian(String v1,String v2){
 		BigDecimal b1 = new BigDecimal(v1);
         BigDecimal b2 = new BigDecimal(v2);
-        return b1.subtract(b2).doubleValue();
+        return b1.subtract(b2);
 	}
 	
-	public static double jian(BigDecimal v1,BigDecimal v2){
-        return v1.subtract(v2).doubleValue();
+	public static BigDecimal jian(BigDecimal v1,BigDecimal v2){
+        return v1.subtract(v2);
 	}
 	
-	public static double cheng(double v1,double v2){
+	public static BigDecimal cheng(double v1,double v2){
 		BigDecimal b1 = new BigDecimal(Double.toString(v1));
         BigDecimal b2 = new BigDecimal(Double.toString(v2));
-        return b1.multiply(b2).doubleValue();
+        return b1.multiply(b2);
 	}
 	
-	public static double cheng(String v1,String v2){
+	public static BigDecimal cheng(String v1,String v2){
 		BigDecimal b1 = new BigDecimal(v1);
         BigDecimal b2 = new BigDecimal(v2);
-        return b1.multiply(b2).doubleValue();
+        return b1.multiply(b2);
 	}
 	
-	public static double cheng(BigDecimal v1,BigDecimal v2){
-        return v1.multiply(v2).doubleValue();
+	public static BigDecimal cheng(BigDecimal v1,BigDecimal v2){
+        return v1.multiply(v2);
 	}
 	
-	public static double chu(double v1,double v2,int scale){
+	public static BigDecimal chu(double v1,double v2,int scale){
 		BigDecimal b1 = new BigDecimal(Double.toString(v1));
         BigDecimal b2 = new BigDecimal(Double.toString(v2));
-        return b1.divide(b2,scale,BigDecimal.ROUND_HALF_UP).doubleValue();
+        return b1.divide(b2,scale,BigDecimal.ROUND_HALF_UP);
 	}
 	
-	public static double chu(double v1,double v2){
+	public static BigDecimal chu(double v1,double v2){
 		BigDecimal b1 = new BigDecimal(Double.toString(v1));
         BigDecimal b2 = new BigDecimal(Double.toString(v2));
-        return b1.divide(b2,10,BigDecimal.ROUND_HALF_UP).doubleValue();
+        return b1.divide(b2,10,BigDecimal.ROUND_HALF_UP);
 	}
 	
-	public static double chu(String v1,String v2,int scale){
+	public static BigDecimal chu(String v1,String v2,int scale){
 		BigDecimal b1 = new BigDecimal(v1);
         BigDecimal b2 = new BigDecimal(v2);
-        return b1.divide(b2,scale,BigDecimal.ROUND_HALF_UP).doubleValue();
+        return b1.divide(b2,scale,BigDecimal.ROUND_HALF_UP);
 	}
 	
-	public static double chu(String v1,String v2){
+	public static BigDecimal chu(String v1,String v2){
 		BigDecimal b1 = new BigDecimal(v1);
         BigDecimal b2 = new BigDecimal(v2);
-        return b1.divide(b2,10,BigDecimal.ROUND_HALF_UP).doubleValue();
+        return b1.divide(b2,10,BigDecimal.ROUND_HALF_UP);
 	}
 	
-	public static double chu(BigDecimal v1,BigDecimal v2){
-        return v1.divide(v2,10,BigDecimal.ROUND_HALF_UP).doubleValue();
+	public static BigDecimal chu(BigDecimal v1,BigDecimal v2){
+        return v1.divide(v2,10,BigDecimal.ROUND_HALF_UP);
 	}
 }
